requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
imageio==2.9.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
networkx==2.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1594919729702/work
opencv-python==4.3.0.36
opt-einsum==3.2.1
pandas==1.0.5
Pillow==7.1.2
protobuf==3.12.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
flask-ngrok==0.0.25